+++ Assets/OxGFrame/AssetLoader/ThirdParty/YooAsset/Editor/AssetBundleBuilder/BuildPipeline/ScriptableBuildPipeline/ScriptableBuildParameters.cs
@@ -15,6 +15,11 @@ namespace YooAsset.Editor
         public ECompressOption CompressOption = ECompressOption.Uncompressed;
 
         /// <summary>
+        /// 从AssetBundle文件头里剥离Unity版本信息
+        /// </summary>
+        public bool StripUnityVersion = false;
+
+        /// <summary>
         /// 禁止写入类型树结构（可以降低包体和内存并提高加载效率）
         /// </summary>
         public bool DisableWriteTypeTree = false;
@@ -69,6 +74,9 @@ namespace YooAsset.Editor
                 buildParams.BundleCompression = UnityEngine.BuildCompression.LZ4;
             else
                 throw new System.NotImplementedException(CompressOption.ToString());
+
+            if (StripUnityVersion)
+                buildParams.ContentBuildFlags |= UnityEditor.Build.Content.ContentBuildFlags.StripUnityVersion;
 
             if (DisableWriteTypeTree)
                 buildParams.ContentBuildFlags |= UnityEditor.Build.Content.ContentBuildFlags.DisableWriteTypeTree;