+++ Assets/OxGFrame/Hotfixer/Scripts/Runtime/Hotfix/Config/HotfixDllConfig.cs
@@ -1,0 +1,20 @@
+using System;
+using System.Collections.Generic;
+
+namespace OxGFrame.Hotfixer
+{
+    [Serializable]
+    public class HotfixDllConfig
+    {
+        public List<string> aotDlls = new List<string>();
+        public List<string> hotfixDlls = new List<string>();
+
+        public HotfixDllConfig() { }
+
+        public HotfixDllConfig(List<string> aotDlls, List<string> hotfixDlls)
+        {
+            this.aotDlls = aotDlls;
+            this.hotfixDlls = hotfixDlls;
+        }
+    }
+}