+++ Assets/OxGFrame/Hotfixer/Scripts/Runtime/Hotfix/Enum/ConfigFileType.cs
@@ -1,0 +1,8 @@
+namespace OxGFrame.Hotfixer
+{
+    public enum ConfigFileType
+    {
+        Json,
+        Bytes
+    }
+}